made course slug unique

+++ mindjunkies/courses/views.py
@@ -282,6 +282,10 @@ class DeleteCourseView(LoginRequiredMixin, TemplateView):
 
     def get(self, request, *args, **kwargs):
         course = get_object_or_404(Course, slug=self.kwargs["course_slug"])
+        if request.user != course.teacher:
+            messages.error(request, "You are not authorized to delete this course.")
+            return redirect(reverse("course_details", kwargs={"slug": course.slug}))
         course.delete()
         messages.success(request, "Course deleted successfully!")
         return redirect(self.success_url)
+
